REFACTOR: Seperate payment services into own package
