+++ src/main/java/de/hhu/propra/sharingplatform/faker/DataFaker.java
@@ -7,8 +7,8 @@ import de.hhu.propra.sharingplatform.dao.UserRepo;
 import de.hhu.propra.sharingplatform.model.Item;
 import de.hhu.propra.sharingplatform.model.Offer;
 import de.hhu.propra.sharingplatform.model.User;
-import de.hhu.propra.sharingplatform.service.ApiService;
-import de.hhu.propra.sharingplatform.service.IPaymentApi;
+import de.hhu.propra.sharingplatform.service.Payment.ApiService;
+import de.hhu.propra.sharingplatform.service.Payment.IPaymentApi;
 import de.hhu.propra.sharingplatform.service.OfferService;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
