+++ src/main/java/de/hhu/propra/sharingplatform/service/ContractService.java
@@ -6,6 +6,7 @@ import de.hhu.propra.sharingplatform.dto.Status;
 import de.hhu.propra.sharingplatform.model.Conflict;
 import de.hhu.propra.sharingplatform.model.Contract;
 import de.hhu.propra.sharingplatform.model.Offer;
+import de.hhu.propra.sharingplatform.service.Payment.IPaymentService;
 import lombok.Data;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
