+++ src/main/java/de/hhu/propra/sharingplatform/service/Payment/IBankAccountService.java
@@ -1,4 +1,4 @@
-package de.hhu.propra.sharingplatform.service;
+package de.hhu.propra.sharingplatform.service.Payment;
 
 import org.springframework.stereotype.Component;
