+++ src/test/java/de/hhu/propra/sharingplatform/service/OfferServiceTest.java
@@ -15,6 +15,9 @@ import de.hhu.propra.sharingplatform.model.Item;
 import de.hhu.propra.sharingplatform.model.Offer;
 import de.hhu.propra.sharingplatform.model.User;
 import java.time.LocalDateTime;
+
+import de.hhu.propra.sharingplatform.service.Payment.ApiService;
+import de.hhu.propra.sharingplatform.service.Payment.PaymentService;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
